test(update): refactor screenobjects and added new tests
